// src/app/web/api/userProfile.ts
export interface UserProfile {
    user_profile_id: number;
    user_name: string;
    first_name: string;
    last_name: string;
    email: string;
    mobile: string;
    period: string;
    startdate: string;
    enddate: string;
    last_login: string;
    user_status: string;
    create_by: string;
    create_date: string;
    update_by: string;
    update_date: string;
    session_id: string;
    actors:string;
}